# perf: introduce `assume` for generating better code