--- a/rust/lance-index/src/vector/pq/distance.rs
+++ b/rust/lance-index/src/vector/pq/distance.rs
@@ -130,8 +130,8 @@ pub(super) fn compute_pq_distance(
     for (sub_vec_idx, vec_indices) in code.chunks_exact(num_vectors).enumerate() {
         let dist_table =
             &distance_table[sub_vec_idx * NUM_CENTROIDS..(sub_vec_idx + 1) * NUM_CENTROIDS];
-        // assume_eq!(dist_table.len(), NUM_CENTROIDS);
-        debug_assert_eq!(vec_indices.len(), distances.len());
+        assume_eq!(dist_table.len(), NUM_CENTROIDS);
+        assume_eq!(vec_indices.len(), distances.len());
         vec_indices
             .iter()
             .zip(distances.iter_mut())