```
construct_dist_table: l2,PQ=96,DIM=1536
                        time:   [116.87 µs 117.06 µs 117.32 µs]
                        change: [-3.9662% -3.5804% -3.1590%] (p = 0.00 < 0.10)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

construct_dist_table: dot,PQ=96,DIM=1536
                        time:   [142.57 µs 142.76 µs 143.20 µs]
                        change: [-3.9276% -3.4103% -2.7466%] (p = 0.00 < 0.10)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

compute_distances: 16000,l2,PQ=96,DIM=1536
                        time:   [370.34 µs 370.98 µs 371.97 µs]
                        change: [-2.0826% -1.6790% -1.3494%] (p = 0.00 < 0.10)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

compute_distances: 16000,cosine,PQ=96,DIM=1536
                        time:   [390.10 µs 391.95 µs 393.27 µs]
                        change: [-4.2590% -2.2274% -0.8685%] (p = 0.02 < 0.10)
                        Change within noise threshold.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild

compute_distances: 16000,dot,PQ=96,DIM=1536
                        time:   [391.99 µs 392.78 µs 393.77 µs]
                        change: [-2.2872% -1.8801% -1.4425%] (p = 0.00 < 0.10)
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
```